+++ apps/api/src/listing/listing.module.ts
@@ -2,17 +2,18 @@ import { Module } from '@nestjs/common';
 import { AdminListingController } from './admin-listing.controller';
 import { HostListingController } from './host-listing.controller';
 import { ListingService } from './listing.service';
+import { AvailabilityService } from './availability.service';
 import { PublicListingController } from './public-listing.controller';
 import { NotificationModule } from '../notification/notification.module';
 
 @Module({
   imports: [NotificationModule],
-  providers: [ListingService],
+  providers: [ListingService, AvailabilityService],
   controllers: [
     HostListingController,
     AdminListingController,
     PublicListingController,
   ],
-  exports: [ListingService],
+  exports: [ListingService, AvailabilityService],
 })
 export class ListingModule {}